Move flickr.rb to services folder.
