+++ .gitignore
@@ -11,6 +11,9 @@
 /db/*.sqlite3
 /db/*.sqlite3-journal
 
+# Ignore flickr.com keys
+/config/flickr.yml
+
 # Ignore all logfiles and tempfiles.
 /log/*.log
 /tmp
