--- config/flickr.yml
@@ -1,4 +1,0 @@
-api_key: b6f69126c6461b943c5d5bedded6f653
-shared_secret: ffe0ae02801a7e05
-access_token: 72157670489056123-f98edfd45e830795
-access_secret: 7c6d42751497f8cf
